+++ implementations/rust/ockam/ockam_command/src/tcp/listener/delete.rs
@@ -1,3 +1,4 @@
+use anyhow::anyhow;
 use clap::Args;
 use colorful::Colorful;
 
@@ -9,6 +10,7 @@ use ockam_core::api::Request;
 use crate::node::{get_node_name, initialize_node_if_default};
 use crate::util::{node_rpc, Rpc};
 use crate::{docs, node::NodeOpts, CommandGlobalOpts};
+use crate::terminal::ConfirmResult;
 
 const AFTER_LONG_HELP: &str = include_str!("./static/delete/after_long_help.txt");
 
@@ -21,6 +23,10 @@ pub struct DeleteCommand {
 
     /// Tcp Listener ID
     pub address: String,
+
+    /// Confirm the deletion without prompting
+    #[arg(display_order = 901, long, short)]
+    yes: bool,
 }
 
 impl DeleteCommand {
@@ -36,22 +42,52 @@ async fn run_impl(
 ) -> crate::Result<()> {
     let node_name = get_node_name(&opts.state, &cmd.node_opts.at_node);
     let node = parse_node_name(&node_name)?;
-    let mut rpc = Rpc::background(&ctx, &opts, &node)?;
-    let req = Request::delete("/node/tcp/listener")
-        .body(models::transport::DeleteTransport::new(cmd.address.clone()));
-    rpc.request(req).await?;
-    rpc.is_ok()?;
+    if cmd.yes {
+        let mut rpc = Rpc::background(&ctx, &opts, &node)?;
+        let req = Request::delete("/node/tcp/listener")
+            .body(models::transport::DeleteTransport::new(cmd.address.clone()));
+        rpc.request(req).await?;
+        rpc.is_ok()?;
 
-    opts.terminal
-        .stdout()
-        .plain(format!(
-            "{} TCP listener with address '{}' has been deleted.",
-            "✔︎".light_green(),
-            &cmd.address
-        ))
-        .machine(&cmd.address)
-        .json(serde_json::json!({ "tcp-listener": { "address": &cmd.address } }))
-        .write_line()?;
+        opts.terminal
+            .stdout()
+            .plain(format!(
+                "{} TCP listener with address '{}' has been deleted.",
+                "✔︎".light_green(),
+                &cmd.address
+            ))
+            .machine(&cmd.address)
+            .json(serde_json::json!({ "tcp-listener": { "address": &cmd.address } }))
+            .write_line()?;
+    } else {
+        match opts.terminal.confirm("This will delete the selected Tcp-listener. Are you sure?")? {
+            ConfirmResult::Yes => {
+                let mut rpc = Rpc::background(&ctx, &opts, &node)?;
+                let req = Request::delete("/node/tcp/listener")
+                    .body(models::transport::DeleteTransport::new(cmd.address.clone()));
+                rpc.request(req).await?;
+                rpc.is_ok()?;
+
+                opts.terminal
+                    .stdout()
+                    .plain(format!(
+                        "{} TCP listener with address '{}' has been deleted.",
+                        "✔︎".light_green(),
+                        &cmd.address
+                    ))
+                    .machine(&cmd.address)
+                    .json(serde_json::json!({ "tcp-listener": { "address": &cmd.address } }))
+                    .write_line()?;
+            }
+            ConfirmResult::No => {
+                return Ok(());
+            }
+            ConfirmResult::NonTTY => {
+                return Err(anyhow!("Use --yes to confirm").into());
+            }
+        }
+    }
+
 
     Ok(())
 }
